+++ src/components/AddPaymentMethodMenu.tsx
@@ -6,7 +6,7 @@ import type {OnyxEntry} from 'react-native-onyx';
 import {useMemoizedLazyExpensifyIcons} from '@hooks/useLazyAsset';
 import useLocalize from '@hooks/useLocalize';
 import useOnyx from '@hooks/useOnyx';
-import {completePaymentOnboarding} from '@libs/actions/IOU';
+import {completePaymentOnboarding} from '@libs/actions/IOU/PayMoneyRequest';
 import {hasRequestFromCurrentAccount} from '@libs/ReportActionsUtils';
 import {isExpenseReport, isIOUReport} from '@libs/ReportUtils';
 import CONST from '@src/CONST';
